api/email/requirements.txt
numpy>=1.20.3
pandas
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
packaging==21.3
pika==1.3.1
pip==20.0.2
pydub==0.25.1
pyparsing==3.0.9
requests==2.28.1
setuptools==44.0.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
mysql-connector-python==8.0.29
Flask_Cors==3.0.10
kafka-python==2.0.2
DateTime==4.7
pybase64==1.2.3
secure-smtplib==0.1.1